feat(frontend): extend exchange worker with ERC4626 tokens
